+++ src/OpenLoco/src/Windows/VehicleList.cpp
@@ -15,6 +15,7 @@
 #include "OpenLoco.h"
 #include "Ui/Dropdown.h"
 #include "Ui/WindowManager.h"
+#include "Vehicles/OrderManager.h"
 #include "Vehicles/Orders.h"
 #include "Vehicles/Vehicle.h"
 #include "Vehicles/VehicleManager.h"
